test(problem-deploy): split backend stack test by resource type
problem-deploy-backend-stack.test.ts was 781 lines / 39 it / 91 real
assertions. Diff readability was poor when a single CFn resource
property changed. Split by CDK resource type into 8 focused files
(<210 lines each) backed by a shared test-helpers module that
memoizes synthDefault() so each worker only synths the stack once.

Closes #1249
