--- infrastructure/test/problem-deploy-backend-stack.test.ts
@@ -1,781 +1,0 @@
-import * as cdk from "aws-cdk-lib";
-import { Match, Template } from "aws-cdk-lib/assertions";
-import { describe, expect, it } from "vitest";
-import { ParticipantPortalLambda } from "../lib/problem-deploy/participant-portal-lambda";
-import { ProblemDeployBackendStack } from "../lib/problem-deploy/problem-deploy-backend-stack";
-
-// 全 it() で同じ Template を使い回す。stack 構造は default props で固定なので、
-// describe ブロック単位で 1 度 synth すれば 13 回 → 1 回に圧縮できる。
-function synthDefault(): Template {
-  const app = new cdk.App();
-  const stack = new ProblemDeployBackendStack(app, "TestStack", {
-    eventBusArn: "arn:aws:events:ap-northeast-1:123456789012:event-bus/test-bus",
-    sourceBucketName: "test-source-bucket",
-    sourceObjectKey: "source.zip",
-    problemsCatalog: {
-      "hello-world": "problems/challenges/hello-world",
-    },
-    problemsScoring: {},
-    problemsEndpoints: {},
-    environmentName: "development",
-  });
-  return Template.fromStack(stack);
-}
-
-function synthWithDeployConcurrentBuildLimit(): Template {
-  const app = new cdk.App();
-  const stack = new ProblemDeployBackendStack(app, "TestStackWithLimit", {
-    eventBusArn: "arn:aws:events:ap-northeast-1:123456789012:event-bus/test-bus",
-    sourceBucketName: "test-source-bucket",
-    sourceObjectKey: "source.zip",
-    problemsCatalog: { "hello-world": "problems/challenges/hello-world" },
-    problemsScoring: {},
-    problemsEndpoints: {},
-    deployConcurrentBuildLimit: 200,
-    environmentName: "development",
-  });
-  return Template.fromStack(stack);
-}
-
-describe("ProblemDeployBackendStack (MVP-1)", () => {
-  const tpl = synthDefault();
-
-  describe("Deployments DDB table", () => {
-    it("should provision 7 DDB tables (Deployments / Events / Teams / CompetitorAccounts / ProblemEndpoints / Disruptions / AdminAuditLog) each with PK/SK and PROVISIONED 1/1", () => {
-      // ADR-004 Phase 1 で Events / Teams、Issue #459 / ADR-002 Phase 2.1 で CompetitorAccounts、
-      // ADR-012 Phase 3.A で ProblemEndpoints、 Issue #888 で Disruptions (Red Team audit + idempotency)、
-      // Issue #950 (ADR-020 Phase D) で AdminAuditLog (admin 操作監査)。
-      // 7 Table すべて DynamoDbLowCapacity Aspect で 1/1 PROVISIONED に均される。
-      tpl.resourceCountIs("AWS::DynamoDB::Table", 7);
-      tpl.hasResourceProperties(
-        "AWS::DynamoDB::Table",
-        Match.objectLike({
-          ProvisionedThroughput: Match.objectLike({
-            ReadCapacityUnits: 1,
-            WriteCapacityUnits: 1,
-          }),
-          KeySchema: Match.arrayWith([
-            Match.objectLike({ AttributeName: "PK", KeyType: "HASH" }),
-            Match.objectLike({ AttributeName: "SK", KeyType: "RANGE" }),
-          ]),
-        }),
-      );
-    });
-
-    it("should enable TTL on expiresAt", () => {
-      tpl.hasResourceProperties(
-        "AWS::DynamoDB::Table",
-        Match.objectLike({
-          TimeToLiveSpecification: Match.objectLike({
-            AttributeName: "expiresAt",
-            Enabled: true,
-          }),
-        }),
-      );
-    });
-  });
-
-  describe("Deploy API Lambda (invoked from tenant API)", () => {
-    it("should run on Node.js 22 / arm64 with BATTLE_PROBLEMS_CATALOG env", () => {
-      tpl.hasResourceProperties(
-        "AWS::Lambda::Function",
-        Match.objectLike({
-          Runtime: "nodejs22.x",
-          Architectures: ["arm64"],
-          Environment: Match.objectLike({
-            Variables: Match.objectLike({
-              BATTLE_PROBLEMS_CATALOG: JSON.stringify({
-                "hello-world": "problems/challenges/hello-world",
-              }),
-            }),
-          }),
-        }),
-      );
-    });
-
-    // #534: CFn StackEvents / StackResources を読む IAM Allow を持つべき
-    // (= deploy job 詳細ページから直接 CFn API を叩くため)。
-    it("should Allow CFn DescribeStackEvents / DescribeStackResources", () => {
-      tpl.hasResourceProperties(
-        "AWS::IAM::Policy",
-        Match.objectLike({
-          PolicyDocument: Match.objectLike({
-            Statement: Match.arrayWith([
-              Match.objectLike({
-                Effect: "Allow",
-                Action: Match.arrayWith([
-                  "cloudformation:DescribeStackEvents",
-                  "cloudformation:DescribeStackResources",
-                ]),
-              }),
-            ]),
-          }),
-        }),
-      );
-    });
-  });
-
-  describe("CodeBuild Project (runs deploy-battles.sh)", () => {
-    it("should create 1 CodeBuild Project", () => {
-      tpl.resourceCountIs("AWS::CodeBuild::Project", 1);
-    });
-
-    it("CodeBuild should read from S3 source", () => {
-      tpl.hasResourceProperties(
-        "AWS::CodeBuild::Project",
-        Match.objectLike({
-          Source: Match.objectLike({
-            Type: "S3",
-            Location: Match.stringLikeRegexp("test-source-bucket/source.zip"),
-          }),
-        }),
-      );
-    });
-
-    it("should not output ConcurrentBuildLimit when `deployConcurrentBuildLimit` is unset (#538)", () => {
-      // 未指定 = AWS account 全体の concurrent build quota (region default 60) を本 Project
-      // でフルに使う既存挙動。本 stack の default props で synth した時点で property が
-      // 出ていないことを保証する (= 既存運用への regression 防止)。
-      const projects = tpl.findResources("AWS::CodeBuild::Project");
-      const project = Object.values(projects)[0] as {
-        Properties?: { ConcurrentBuildLimit?: number };
-      };
-      expect(project?.Properties?.ConcurrentBuildLimit).toBeUndefined();
-    });
-  });
-
-  describe("CodeBuild Project concurrent build limit (#538)", () => {
-    // 共有 fixture (`tpl = synthDefault()`) と別 props を渡すので別 instance で synth する。
-    // bundling が長い環境でも assertion timeout に巻き込まれないよう collection 時に fixture 化する。
-    const limited = synthWithDeployConcurrentBuildLimit();
-
-    it("should reflect `deployConcurrentBuildLimit: 200` in the CFn property", () => {
-      limited.hasResourceProperties(
-        "AWS::CodeBuild::Project",
-        Match.objectLike({ ConcurrentBuildLimit: 200 }),
-      );
-    });
-  });
-
-  describe("Step Functions State Machine + EventBridge Rule", () => {
-    it("should create 3 State Machines (Create / Delete / BulkCreate) (Issue #910 Phase 2.C.2.a)", () => {
-      tpl.resourceCountIs("AWS::StepFunctions::StateMachine", 3);
-    });
-
-    it("Create State Machine should write the PENDING → IN_PROGRESS intermediate transition before starting CodeBuild", () => {
-      // RUN_JOB 同期 CodeBuild は 5〜15 分かかるため、この中間書込が無いと operator UI が
-      // PENDING のまま固定して polling が機能していないように見える (#159 の再発防止)。
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const synthJson = JSON.stringify(stateMachines);
-      expect(synthJson).toContain("MarkInProgress");
-      expect(synthJson).toContain("IN_PROGRESS");
-    });
-
-    it("Create State Machine should persist the CodeBuild buildId into the Deployments row on completion/failure", () => {
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const createStateMachine = Object.values(stateMachines)
-        .map((stateMachine) => JSON.stringify(stateMachine))
-        .find((definition) => definition.includes("StartDeployCodeBuild"));
-
-      expect(createStateMachine).toBeDefined();
-      expect(createStateMachine).toContain("MarkSucceeded");
-      expect(createStateMachine).toContain("MarkFailed");
-      expect(createStateMachine).toContain("PROBLEM_EXTERNAL_ID");
-      expect(createStateMachine).toContain("$.detail.jobId");
-      expect(createStateMachine).toContain(
-        "stackId = :stackId, stackOutputs = :stackOutputs, buildId = :buildId",
-      );
-      expect(createStateMachine).toContain("#failureReason = :failureReason, buildId = :buildId");
-      expect(createStateMachine).toContain("$.codebuild.Build.Id");
-    });
-
-    it("Create State Machine should Catch CodeBuild timeout / AccessDenied and fall to FAILED", () => {
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const createStateMachine = Object.values(stateMachines)
-        .map((stateMachine) => JSON.stringify(stateMachine))
-        .find((definition) => definition.includes("StartDeployCodeBuild"));
-
-      expect(createStateMachine).toBeDefined();
-      expect(createStateMachine).toContain("StartDeployCodeBuild");
-      expect(createStateMachine).toContain("StartDeployCodeBuildCrossAccount");
-      expect(createStateMachine).toContain("States.ALL");
-      expect(createStateMachine).toContain("RouteFailedDeployment");
-      expect(createStateMachine).toContain("MarkFailed");
-      expect(createStateMachine).toContain("MarkFailedWithoutBuildId");
-    });
-
-    it("Create State Machine should treat ROLLBACK_COMPLETE as terminal failure and fall to MarkFailed", () => {
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const createStateMachine = Object.values(stateMachines)
-        .map((stateMachine) => JSON.stringify(stateMachine))
-        .find((definition) => definition.includes("StartDeployCodeBuild"));
-
-      expect(createStateMachine).toBeDefined();
-      expect(createStateMachine).toContain("RouteDescribedStackStatus");
-      expect(createStateMachine).toContain("UseStackStatusReasonAsFailureCause");
-      expect(createStateMachine).toContain("$.cfn.Stacks[0].StackStatus");
-      expect(createStateMachine).toContain("ROLLBACK_COMPLETE");
-      expect(createStateMachine).toContain("CREATE_FAILED");
-      expect(createStateMachine).toContain("UPDATE_ROLLBACK_COMPLETE");
-      expect(createStateMachine).toContain("$.cfn.Stacks[0].StackStatusReason");
-      expect(createStateMachine).toContain("MarkFailed");
-    });
-
-    it("Create State Machine should run DescribeStacks via Lambda (#762)", () => {
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const createStateMachine = Object.values(stateMachines)
-        .map((stateMachine) => JSON.stringify(stateMachine))
-        .find((definition) => definition.includes("StartDeployCodeBuild"));
-
-      expect(createStateMachine).toBeDefined();
-      expect(createStateMachine).toContain("RouteCreateInput");
-      expect(createStateMachine).toContain("StartDeployCodeBuildCrossAccount");
-      expect(createStateMachine).toContain("DescribeStack");
-      expect(createStateMachine).toContain("DescribeStackFunction");
-      expect(createStateMachine).not.toContain(
-        "arn:aws:states:::aws-sdk:cloudformation:describeStacks",
-      );
-      expect(createStateMachine).not.toContain("States.Format('{}', $.detail.competitorRoleArn)");
-    });
-
-    it("Delete State Machine should Choice on the presence of AssumeRole metadata and not die at runtime on missing path references (#758)", () => {
-      const stateMachines = tpl.findResources("AWS::StepFunctions::StateMachine");
-      const deleteStateMachine = Object.values(stateMachines)
-        .map((stateMachine) => JSON.stringify(stateMachine))
-        .find((definition) => definition.includes("StartDeleteCodeBuild"));
-
-      expect(deleteStateMachine).toBeDefined();
-      expect(deleteStateMachine).toContain("RouteDeleteInput");
-      expect(deleteStateMachine).toContain("StartDeleteCodeBuildCrossAccount");
-      expect(deleteStateMachine).toContain("InvalidAssumeRoleMetadata");
-      expect(deleteStateMachine).toContain("$.detail.competitorRoleArn");
-      expect(deleteStateMachine).toContain("$.detail.externalIdParameterName");
-      expect(deleteStateMachine).toContain("MarkFailed");
-    });
-
-    it("should have 7 EventBridge Rules (Create / Delete / BulkCreate / GenericScoring / ExternalIdAudit schedule / SystemAuditWriter (Issue #1034) / CodeBuildFailure (Issue #1029))", () => {
-      // 旧 2 (Create / Delete state-machine event rules)
-      //   + BulkCreate (Issue #910 Phase 2.C: BulkDeployCreateRequested → Distributed Map)
-      //   + GenericScoring schedule rate(1 minute) (= ADR-012 Phase 3.B、 旧 HealthCheck 後継)
-      //   + ExternalIdAudit schedule rate(1 day) (= Phase 3.2 / Issue #603 で追加)
-      // = 5。GenericScoring は scoring 問題が無い tenant でも reconcile 用に常時 instantiate される。
-      // 旧 5 + Issue #1034 SystemAuditWriter (SBT bus) + Issue #1029 CodeBuildFailure (default bus)
-      tpl.resourceCountIs("AWS::Events::Rule", 7);
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          EventPattern: Match.objectLike({
-            source: ["tenkacloud.deploy"],
-            "detail-type": ["DeployCreateRequested"],
-          }),
-        }),
-      );
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          EventPattern: Match.objectLike({
-            source: ["tenkacloud.deploy"],
-            "detail-type": ["DeployDeleteRequested"],
-          }),
-        }),
-      );
-      // GenericScoring の rate(1 minute) schedule (= ADR-012 Phase 3.B dispatcher + reconciler)
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          ScheduleExpression: "rate(1 minute)",
-        }),
-      );
-      // ExternalIdAudit の rate(1 day) schedule (Phase 3.2 / Issue #603)
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          ScheduleExpression: "rate(1 day)",
-        }),
-      );
-    });
-
-    it("GenericScoring Lambda should have table-name env without the catalog env", () => {
-      // Issue #1158: BATTLE_PROBLEMS_SCORING / PROBLEM_ENDPOINTS / BATTLE_PROBLEMS_PHASES は
-      // env vars 4 KB 上限を回避するため esbuild bundling.define で build 時 literal 置換し、
-      // Lambda env からは取り除いている。 catalog 配線は handler の `process.env.X` 読み取りが
-      // build 時に literal JSON に固定され、 runtime IO 0 で動く。
-      const functions = tpl.findResources("AWS::Lambda::Function");
-      const genericScoring = Object.entries(functions).find(
-        ([name]) => name.includes("GenericScoring") && name.includes("Function"),
-      );
-      expect(genericScoring).toBeDefined();
-      const vars =
-        (
-          genericScoring?.[1] as {
-            Properties?: { Environment?: { Variables?: Record<string, unknown> } };
-          }
-        )?.Properties?.Environment?.Variables ?? {};
-      expect(vars.DEPLOYMENTS_TABLE_NAME).toBeDefined();
-      expect(vars.EVENTS_TABLE_NAME).toBeDefined();
-      expect(vars.PROBLEM_ENDPOINTS_TABLE_NAME).toBeDefined();
-      expect(vars.BATTLE_PROBLEMS_SCORING).toBeUndefined();
-      expect(vars.PROBLEM_ENDPOINTS).toBeUndefined();
-      expect(vars.BATTLE_PROBLEMS_PHASES).toBeUndefined();
-    });
-  });
-
-  describe("Outputs", () => {
-    it("should expose DeploymentsTableName and DeployCreateStateMachineArn as Outputs", () => {
-      const outputs = tpl.findOutputs("*");
-      expect(Object.keys(outputs)).toEqual(
-        expect.arrayContaining(["DeploymentsTableName", "DeployCreateStateMachineArn"]),
-      );
-    });
-
-    it("ADR-012 Phase 3.A: should expose ProblemEndpointsTableName as an Output", () => {
-      const outputs = tpl.findOutputs("*");
-      expect(Object.keys(outputs)).toEqual(expect.arrayContaining(["ProblemEndpointsTableName"]));
-    });
-
-    it("Issue #1053: should expose CompetitorBootstrapTemplateUrl as an Output", () => {
-      // hosting を AdminConsoleHostingStack から移管したため、 本 stack が出力 owner。
-      // SaaS の AdminConsoleHosting と Lite / SaaS の ApplicationAdminConsoleHosting が
-      // cross-stack ref で受け取る。
-      const outputs = tpl.findOutputs("*");
-      expect(Object.keys(outputs)).toEqual(
-        expect.arrayContaining(["CompetitorBootstrapTemplateUrl"]),
-      );
-    });
-  });
-
-  describe("Issue #1053: CompetitorBootstrapHosting (公開 S3 hosting)", () => {
-    it("should create 1 additional public-read S3 bucket (migrated from the old AdminConsoleHostingStack)", () => {
-      // ParticipantPortal 等で別 bucket もあるため count assert は避け、 public-read 設定の
-      // ある bucket が存在することで pin。
-      tpl.hasResourceProperties(
-        "AWS::S3::Bucket",
-        Match.objectLike({
-          PublicAccessBlockConfiguration: Match.objectLike({
-            BlockPublicAcls: false,
-            BlockPublicPolicy: false,
-            IgnorePublicAcls: false,
-            RestrictPublicBuckets: false,
-          }),
-        }),
-      );
-    });
-  });
-
-  describe("Issue #1034: SystemAuditWriter Lambda + EventBridge Rule", () => {
-    it("should have an EventBridge Rule listening to the 6 SBT onboarding/offboarding detailTypes", () => {
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          EventPattern: Match.objectLike({
-            "detail-type": Match.arrayWith([
-              "onboardingRequest",
-              "onboardingSuccess",
-              "onboardingFailure",
-              "offboardingRequest",
-              "offboardingSuccess",
-              "offboardingFailure",
-            ]),
-          }),
-        }),
-      );
-    });
-
-    it("SystemAuditWriter Lambda should have DEPLOY_ENVIRONMENT + ADMIN_AUDIT_LOG_TABLE_NAME env", () => {
-      // env が無いと writeAuditEvent が no-op になり、 audit が書かれない silent failure に戻る。
-      tpl.hasResourceProperties(
-        "AWS::Lambda::Function",
-        Match.objectLike({
-          Runtime: "nodejs22.x",
-          Architectures: ["arm64"],
-          Environment: Match.objectLike({
-            Variables: Match.objectLike({
-              ADMIN_AUDIT_LOG_TABLE_NAME: Match.anyValue(),
-              DEPLOY_ENVIRONMENT: "development",
-            }),
-          }),
-        }),
-      );
-    });
-
-    it("Issue #1029: should have a Rule catching CodeBuild FAILED / FAULT / STOPPED / TIMED_OUT events", () => {
-      tpl.hasResourceProperties(
-        "AWS::Events::Rule",
-        Match.objectLike({
-          EventPattern: Match.objectLike({
-            source: ["aws.codebuild"],
-            "detail-type": ["CodeBuild Build State Change"],
-            detail: Match.objectLike({
-              "build-status": Match.arrayWith(["FAILED", "FAULT", "STOPPED", "TIMED_OUT"]),
-            }),
-          }),
-        }),
-      );
-    });
-  });
-
-  describe("legacy 経路の廃止", () => {
-    it("should not create the legacy DeployApiGateway (HTTP API)", () => {
-      tpl.resourceCountIs("AWS::ApiGatewayV2::Api", 0);
-    });
-  });
-
-  describe("Competitor Accounts API Lambda (Issue #459 / ADR-002 Phase 2.1)", () => {
-    it("Lambda should have COMPETITOR_ACCOUNTS_TABLE_NAME / DEPLOY_ENVIRONMENT / TENKACLOUD_ACCOUNT_ID env", () => {
-      tpl.hasResourceProperties(
-        "AWS::Lambda::Function",
-        Match.objectLike({
-          Runtime: "nodejs22.x",
-          Architectures: ["arm64"],
-          Environment: Match.objectLike({
-            Variables: Match.objectLike({
-              COMPETITOR_ACCOUNTS_TABLE_NAME: Match.anyValue(),
-              DEPLOY_ENVIRONMENT: "development",
-              TENKACLOUD_ACCOUNT_ID: Match.anyValue(),
-            }),
-          }),
-        }),
-      );
-    });
-
-    it("should grant least-privilege SSM Parameter Store + STS AssumeRole to the Lambda Role", () => {
-      // SSM は path prefix で絞り込み、STS は TenkaCloud- Role 名 pattern で絞り込み。
-      tpl.hasResourceProperties(
-        "AWS::IAM::Policy",
-        Match.objectLike({
-          PolicyDocument: Match.objectLike({
-            Statement: Match.arrayWith([
-              Match.objectLike({
-                Effect: "Allow",
-                Action: Match.arrayWith(["ssm:GetParameter", "ssm:PutParameter"]),
-              }),
-            ]),
-          }),
-        }),
-      );
-      tpl.hasResourceProperties(
-        "AWS::IAM::Policy",
-        Match.objectLike({
-          PolicyDocument: Match.objectLike({
-            Statement: Match.arrayWith([
-              Match.objectLike({
-                Effect: "Allow",
-                Action: "sts:AssumeRole",
-                Resource: "arn:aws:iam::*:role/TenkaCloud-*",
-              }),
-            ]),
-          }),
-        }),
-      );
-    });
-
-    it("should include CompetitorAccountsTableName in Outputs", () => {
-      const outputs = tpl.findOutputs("*");
-      expect(Object.keys(outputs)).toEqual(expect.arrayContaining(["CompetitorAccountsTableName"]));
-    });
-
-    it("Phase 3.2 / Issue #603: ExternalIdAudit Lambda should run on rate(1 day) and have PutMetricData (namespace-scoped)", () => {
-      // rate(1 day) schedule は上の resourceCountIs(Rule, 4) の test でも assert 済だが、ここでは
-      // ExternalIdAudit 経路の代表 evidence (COMPETITOR_ACCOUNTS_TABLE_NAME env + namespace 絞り込み IAM) を pin する。
-      tpl.hasResourceProperties(
-        "AWS::Lambda::Function",
-        Match.objectLike({
-          Runtime: "nodejs22.x",
-          Environment: Match.objectLike({
-            Variables: Match.objectLike({
-              COMPETITOR_ACCOUNTS_TABLE_NAME: Match.anyValue(),
-              DEPLOY_ENVIRONMENT: "development",
-            }),
-          }),
-        }),
-      );
-      tpl.hasResourceProperties(
-        "AWS::IAM::Policy",
-        Match.objectLike({
-          PolicyDocument: Match.objectLike({
-            Statement: Match.arrayWith([
-              Match.objectLike({
-                Effect: "Allow",
-                Action: "cloudwatch:PutMetricData",
-                Resource: "*",
-                Condition: Match.objectLike({
-                  StringEquals: Match.objectLike({
-                    "cloudwatch:namespace": "TenkaCloud/CompetitorAccounts",
-                  }),
-                }),
-              }),
-            ]),
-          }),
-        }),
-      );
-    });
-
-    it("KMS policy should have StringLike + Decrypt/GenerateDataKey (supports both SSM SecureString GET/PUT)", () => {
-      // **regression 防止**: 旧実装は StringEquals + Decrypt/Encrypt の組合せだった (PR-594
-      // /security-review Vuln 1)。`StringEquals` は wildcard 展開しないので Condition が
-      // 永久に false で fail-closed、Encrypt は SecureString PUT に不適合 (GenerateDataKey 必要)。
-      // この test で StringLike + Decrypt + GenerateDataKey の組合せを pin する。
-      tpl.hasResourceProperties(
-        "AWS::IAM::Policy",
-        Match.objectLike({
-          PolicyDocument: Match.objectLike({
-            Statement: Match.arrayWith([
-              Match.objectLike({
-                Effect: "Allow",
-                Action: Match.arrayWith(["kms:Decrypt", "kms:GenerateDataKey"]),
-                Resource: "*",
-                Condition: Match.objectLike({
-                  StringLike: Match.objectLike({
-                    "kms:EncryptionContext:PARAMETER_ARN": Match.anyValue(),
-                  }),
-                }),
-              }),
-            ]),
-          }),
-        }),
-      );
-    });
-  });
-});
-
-/**
- * ParticipantPortalLambda 単体 synth (#535 再発防止)。
- *
- * Stack 全体を synth すると `ParticipantPortalHosting` が
- * `apps/participant-portal/dist` の asset を要求し、CI 環境 (= dist 未 build) で
- * fail する。Lambda の env / IAM だけ確認できれば十分なので、Lambda construct を
- * 単体で synth する。
- */
-function synthParticipantPortalLambdaOnly(): Template {
-  const app = new cdk.App();
-  const stack = new cdk.Stack(app, "TestStack");
-  const deployments = new cdk.aws_dynamodb.Table(stack, "Deployments", {
-    partitionKey: { name: "PK", type: cdk.aws_dynamodb.AttributeType.STRING },
-    sortKey: { name: "SK", type: cdk.aws_dynamodb.AttributeType.STRING },
-  });
-  const events = new cdk.aws_dynamodb.Table(stack, "Events", {
-    partitionKey: { name: "PK", type: cdk.aws_dynamodb.AttributeType.STRING },
-    sortKey: { name: "SK", type: cdk.aws_dynamodb.AttributeType.STRING },
-  });
-  const endpoints = new cdk.aws_dynamodb.Table(stack, "ProblemEndpoints", {
-    partitionKey: { name: "PK", type: cdk.aws_dynamodb.AttributeType.STRING },
-    sortKey: { name: "SK", type: cdk.aws_dynamodb.AttributeType.STRING },
-  });
-  new ParticipantPortalLambda(stack, "ParticipantPortal", {
-    deploymentsTable: deployments,
-    eventsTable: events,
-    endpointsTable: endpoints,
-    problemsScoring: {},
-    problemsEndpoints: {},
-    environmentName: "development",
-  });
-  return Template.fromStack(stack);
-}
-
-describe("ParticipantPortalLambda wiring (#535)", () => {
-  const tpl = synthParticipantPortalLambdaOnly();
-
-  it("should set EVENTS_TABLE_NAME in the ParticipantPortal Lambda environment", () => {
-    // ADR-006 Notifications backend (PR-524) が Module load 時に EVENTS_TABLE_NAME を
-    // 必須で読むので、CDK 配線が無いと Lambda init で throw して portal 全 route が
-    // 502 になる (= #535 regression)。本 assertion で再発防止。
-    tpl.hasResourceProperties(
-      "AWS::Lambda::Function",
-      Match.objectLike({
-        Environment: Match.objectLike({
-          Variables: Match.objectLike({
-            DEPLOYMENTS_TABLE_NAME: Match.anyValue(),
-            EVENTS_TABLE_NAME: Match.anyValue(),
-          }),
-        }),
-      }),
-    );
-  });
-
-  it("should grant Events-table dynamodb:Query + GetItem to the ParticipantPortal Lambda IAM Role", () => {
-    // ADR-006: GET /portal/me/notifications が Events table を Query する (= partition 単位)。
-    // Issue #1005: submit-flag / hint reveal が共有する event-gate.ts が PK=EVENT#<id> /
-    // SK=META を `dynamodb:GetItem` で 1 行引く (= scoring gate)。 grant が漏れていると
-    // AccessDenied で getEventGate が undefined を返し、 fail-closed で `scoring_not_started`
-    // に倒れ 「Event 採点中なのに flag 提出が reject される」 不整合になる (= 実 deploy で
-    // 観測した regression、 CloudWatch logs `[event-gate] getEventGate failed AccessDenied`)。
-    // 配線が無いと AccessDenied で 500 / 提出 reject になる。 Role 直貼りの inline policy
-    // なので `AWS::IAM::Role` の Policies 配列を見る。
-    tpl.hasResourceProperties(
-      "AWS::IAM::Role",
-      Match.objectLike({
-        Policies: Match.arrayWith([
-          Match.objectLike({
-            PolicyName: "EventsRead",
-            PolicyDocument: Match.objectLike({
-              Statement: Match.arrayWith([
-                Match.objectLike({
-                  Action: Match.arrayWith(["dynamodb:Query", "dynamodb:GetItem"]),
-                  Effect: "Allow",
-                }),
-              ]),
-            }),
-          }),
-        ]),
-      }),
-    );
-  });
-
-  it("ADR-012 Phase 3.A: should set PROBLEM_ENDPOINTS_TABLE_NAME env without the catalog env", () => {
-    // Issue #1158: PROBLEM_ENDPOINTS / BATTLE_PROBLEMS_SCORING は env 4 KB 上限回避のため
-    // esbuild bundling.define で build 時 literal 置換し、 Lambda env からは取り除いている。
-    tpl.hasResourceProperties(
-      "AWS::Lambda::Function",
-      Match.objectLike({
-        Environment: Match.objectLike({
-          Variables: Match.objectLike({
-            PROBLEM_ENDPOINTS_TABLE_NAME: Match.anyValue(),
-          }),
-        }),
-      }),
-    );
-    const functions = tpl.findResources("AWS::Lambda::Function");
-    const participantPortal = Object.entries(functions).find(
-      ([name]) => name.includes("ParticipantPortal") && name.includes("Function"),
-    );
-    expect(participantPortal).toBeDefined();
-    const vars =
-      (
-        participantPortal?.[1] as {
-          Properties?: { Environment?: { Variables?: Record<string, unknown> } };
-        }
-      )?.Properties?.Environment?.Variables ?? {};
-    expect(vars.PROBLEM_ENDPOINTS).toBeUndefined();
-    expect(vars.BATTLE_PROBLEMS_SCORING).toBeUndefined();
-  });
-
-  it("should have DEPLOY_ENVIRONMENT for AWS Console SSO", () => {
-    const functions = tpl.findResources("AWS::Lambda::Function");
-    const fn = Object.values(functions)[0] as {
-      Properties?: { Environment?: { Variables?: Record<string, unknown> } };
-    };
-    const vars = fn.Properties?.Environment?.Variables ?? {};
-    expect(vars.DEPLOY_ENVIRONMENT).toBe("development");
-  });
-
-  it("should have SSM ExternalId read and CompetitorDeployRole AssumeRole permissions for AWS Console SSO", () => {
-    tpl.hasResourceProperties(
-      "AWS::IAM::Role",
-      Match.objectLike({
-        Policies: Match.arrayWith([
-          Match.objectLike({
-            PolicyName: "ConsoleSso",
-            PolicyDocument: Match.objectLike({
-              Statement: Match.arrayWith([
-                Match.objectLike({
-                  Action: "ssm:GetParameter",
-                  Effect: "Allow",
-                }),
-                Match.objectLike({
-                  Action: "sts:AssumeRole",
-                  Effect: "Allow",
-                  Resource: "arn:aws:iam::*:role/TenkaCloud-*",
-                }),
-              ]),
-            }),
-          }),
-        ]),
-      }),
-    );
-  });
-
-  it("ADR-012 Phase 3.A: should grant Endpoints table Query / PutItem / DeleteItem to the IAM Role", () => {
-    tpl.hasResourceProperties(
-      "AWS::IAM::Role",
-      Match.objectLike({
-        Policies: Match.arrayWith([
-          Match.objectLike({
-            PolicyName: "EndpointsRW",
-            PolicyDocument: Match.objectLike({
-              Statement: Match.arrayWith([
-                Match.objectLike({
-                  Action: Match.arrayWith([
-                    "dynamodb:Query",
-                    "dynamodb:PutItem",
-                    "dynamodb:DeleteItem",
-                  ]),
-                  Effect: "Allow",
-                }),
-              ]),
-            }),
-          }),
-        ]),
-      }),
-    );
-  });
-});
-
-describe("ProblemDeployBackendStack (#778 ADR-016 Phase 2: eventBusArn optional 化)", () => {
-  // synth は 5 個の NodejsFunction (= esbuild bundling) を含むため CI 上で ~7s かかる。
-  // vitest の default 5s timeout を 30s に拡張する (= 既存 #538 test と同じ pattern)。
-  const SYNTH_TIMEOUT_MS = 30_000;
-
-  // describe scope で 1 度だけ synth して、 3 件の it で再利用 (= per-test の重複 synth で
-  // 21s 消費するのを 7s に圧縮)。
-  let liteTemplate: Template;
-  let fullTemplate: Template;
-
-  it(
-    "should create 1 new local EventBus when eventBusArn is omitted (Lite mode self-contained)",
-    () => {
-      const app = new cdk.App();
-      const stack = new ProblemDeployBackendStack(app, "LiteStack", {
-        sourceBucketName: "test-source-bucket-lite",
-        sourceObjectKey: "source.zip",
-        problemsCatalog: { "hello-world": "problems/challenges/hello-world" },
-        problemsScoring: {},
-        problemsEndpoints: {},
-        environmentName: "development",
-        // eventBusArn を省略 (= Lite mode)
-      });
-      liteTemplate = Template.fromStack(stack);
-      liteTemplate.resourceCountIs("AWS::Events::EventBus", 1);
-      liteTemplate.hasResourceProperties(
-        "AWS::Events::EventBus",
-        Match.objectLike({
-          Name: Match.stringLikeRegexp("^tenkacloud-problem-deploy-local-"),
-        }),
-      );
-    },
-    SYNTH_TIMEOUT_MS,
-  );
-
-  it(
-    "should not create a local EventBus in the existing path (Full mode) where eventBusArn is provided",
-    () => {
-      fullTemplate = synthDefault();
-      fullTemplate.resourceCountIs("AWS::Events::EventBus", 0);
-    },
-    SYNTH_TIMEOUT_MS,
-  );
-
-  it(
-    "should still provision DeployApi / EventApi / CompetitorAccountsApi / GenericScoring Lambdas in the same shape even when eventBusArn is omitted (features stay live)",
-    () => {
-      // 前の test で立てた liteTemplate を再利用 (= synth コストを節約)。
-      if (!liteTemplate) {
-        const app = new cdk.App();
-        const stack = new ProblemDeployBackendStack(app, "LiteStack2", {
-          sourceBucketName: "test-source-bucket-lite-2",
-          sourceObjectKey: "source.zip",
-          problemsCatalog: { "hello-world": "problems/challenges/hello-world" },
-          problemsScoring: {},
-          problemsEndpoints: {},
-          environmentName: "development",
-        });
-        liteTemplate = Template.fromStack(stack);
-      }
-      const lambdaCount = Object.keys(liteTemplate.findResources("AWS::Lambda::Function")).length;
-      expect(lambdaCount).toBeGreaterThan(0);
-      // EventBridge Rule (= Step Functions trigger) も local bus にぶら下がる。
-      const ruleCount = Object.keys(liteTemplate.findResources("AWS::Events::Rule")).length;
-      expect(ruleCount).toBeGreaterThan(0);
-    },
-    SYNTH_TIMEOUT_MS,
-  );
-});
